requirements.txt
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
humanfriendly==10.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
onnx==1.15.0
onnxruntime==1.17.1
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
protobuf==5.26.0
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
segment-anything==1.0
six==1.16.0
sympy==1.12
torch==2.2.1
triton==2.2.0
typing_extensions==4.10.0
